Remove dependency on remix-utils
I ran into a problem where remix-auth depends on an old version of
remix-utils, and this version of remix-utils in turn does not support
React 18. My initial plan was simply to upgrade to the latest version of
remix-utils, but I realized that remix-auth only uses remix-utils for
these response functions. While they certainly look nicer, it feels a
bit unnecessary to pull in an entire dependency just for that, so I
figured it'd be better to just remove the dependency on remix-utils.
This will make it easier to keep remix-auth up to date in the future.